+++ vm/asyncio.go
@@ -293,6 +293,8 @@ func (i *Interp) builtinModule(name string) (*object.Module, bool) {
 		return i.buildAnnotationlib(), true
 	case "site":
 		return i.buildSite(), true
+	case "code":
+		return i.buildCode(), true
 	case "sys.monitoring":
 		return i.buildSysMonitoring(), true
 	case "shutil":